enforce `int64` on all platforms

+++ dask/array/tests/test_routines.py
@@ -1599,7 +1599,7 @@ def test_take():
 
 
 def test_take_large():
-    a = da.arange(1_000_000_000_000, chunks=(200_000_000,))
+    a = da.arange(1_000_000_000_000, chunks=(200_000_000,), dtype="int64")
     x = np.arange(20)
 
     assert_eq(da.take(a, x, axis=0), x)
